Manual sorted residencyProgram.js file: Not applicable, hospital, country
